requirement.txt
asyncio
aiogram
pydantic
pydantic-settings
ffmpeg-python